Extract FontLoaderView
