+++ src/App.tsx
@@ -1,10 +1,13 @@
 import { SafeAreaProvider } from "react-native-safe-area-context";
+import { FontLoaderView } from "./FontLoaderView";
 import { MainView } from "./MainView";
 
 export default function App() {
   return (
     <SafeAreaProvider>
-      <MainView />
+      <FontLoaderView>
+        <MainView />
+      </FontLoaderView>
     </SafeAreaProvider>
   );
 }
